feat: auto fedGLM
