+++ python/primihub/new_FL/tests/example/ChatGlm.json
@@ -12,11 +12,17 @@
                 "port": "50051",
                 "use_tls": false
             },
+        "Charlie":
+            {
+                "ip": "127.0.0.1",
+                "port": "50052",
+                "use_tls": false
+            },
         "task_manager": "127.0.0.1:50050"
     },
     "roles":
-        {"guest":["Alice"],
-        "host": ["Bob"]},
+        {"client":["Alice", "Bob"],
+        "server": ["Charlie"]},
     "tasks":[
         {
         "model": "Chat_glm",
